Upgraded to the following software versions:     Aristotle 2.35.0
Added mention of enabling file type detection for Vim in help message from
sde_ycmconf.
Revised the MOTD message, which was outdated.
Changed all of the previously-existing environments not to use the
SDE-installed version of SPL.
Added a new environment cmake_spl that does use the SDE-installed version
of SPL.