Fixed a bug in the sde_install_cmake script that caused incorrect behavior if a relative path is used to specify the installation directory.
Added a new option to the sde_install_boost script that causes the CMake
configuration files built by Boost not to be used (by installing them
in a directory where they will not normally be found by CMake).
This functionality is sometimes needed to workaround bugs in Boost.

Changed the installer to specify that the installation of Boost should
not use the CMake configuration files that are built as part of Boost.